There is fd leak in sonivox lib when try open any media file > 2GB (MIDI parser memory leak)
Midi Parser support only 32 bit file operations and thus files
of size less than 2 GB. When file open is called on the Midi
parser, for files more than 2GB, internal seek operation fails
resulting in the non-closure of the opened file. This resulted
in the memory leak of the opened file descriptor. This is
fixed by closing the opened file descriptor on failed cases.

Change-Id: Ie9f53275206e2b4616d1cfc41c12b90544895548
Author: Muthukumar Kandasamy <muthukumar.kandasamy@intel.com>
Signed-off-by: Muthukumar Kandasamy <muthukumar.kandasamy@intel.com>
Signed-off-by: Gurudatta Bhakte <gurudattax.bhakte@intel.com>
Singed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 39292