sonivox: prevent rejection of good but large MIDI files
A security patch that was trying to detect infinite loops
was accidentally rejecting very large MIDI files.
This is because there is a scanning pass that looks at the entire
file. That was generating a very high eventCount.
With this change, we do not check event counts during the scanning pass.

Bug: 112735915
Bug: 112575219
Bug: 68664359
Test: Generate a MIDI files with more than 50000 events.
Test: There are some in b/112735915 and b/112575219
Test: mmma frameworks/av/cmds/stagefright
Test: adb push out/target/product/marlin/system/bin/stagefright /system/bin/.
Test: adb shell stagefright -a /sdcard/Download/verybigfile.mid
Test: It should play correctly and not abort.
Change-Id: Iddf2f5b178e9ca3867b14fcd78d538023d79240d
Merged-In: Iddf2f5b178e9ca3867b14fcd78d538023d79240d