libsonivoxwithoutjet is marked as available to platform
The lib is depended on by libmedia_midiiowrapper and which is again
depend on by libmedia_jni, etc. that are installed to the platform.
With b/153073816, we will be checking the availability to the platform,
i.e. platform-available module depending on platform-unavailable module
causes a build error.

Prepare for the change by correctly marking that this interface lib is
available to the platform.

Bug: 153073816
Test: m
Change-Id: Idb335421d234b46684bc38b78058175260e966d0