Fix ARM assembler for clang.
Use the C preprocessor because clang doesn't have the GNU .if and
friends. Also use #if/#elif/#error rather than assume that one or other
of SAMPLES_8_BIT and SAMPLES_16_BIT has been defined.

Remove the .func/.endfunc stuff for emitting STABS debug info, which
we've never used and LLVM doesn't support.

Fix one instance of an instruction (LDRSH) where the condition code
needed to be moved to the end.

Bug: http://b/133391830
Test: treehugger
Change-Id: I5542c8e1b521c71ed6bcbfa9df84aef1330e1925