Add diagnostics for diagnosing failures
we've started seeing flakiness in the Sonivox testing, which has the
appearance of being caused by the test data not being populated onto
the device.

This adds some diagnostics to the test output so that we see what is in
the test asset directory on the device.

Bug: 384791354
Flag: TEST_ONLY
Test: atest SonivoxTest
Change-Id: I6485281e9536ed7f92f66fc450e64584704a4036